pyproject.toml
[tool.poetry]
name = "design-10x-fixed-probes"
version = "0.1.0"
description = "Design probes for the 10x Genomics Fixed RNA assay"
authors = ["Jay Hesselberth <jay.hesselberth@cuanschutz.edu>"]
license = "MIT"
readme = "README.md"
packages = [{include = "design_10x_fixed_probes"}]

[tool.poetry.dependencies]
python = "^3.10"
click = "^8.1.3"
python-rle = "^0.0.3"

[tool.poetry.group.dev.dependencies]
biopython = "^1.80"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
